Updated Posts: encapsulated styles.
