+++ src/scenes/Home/components/Posts/Posts.js
@@ -1,39 +1,11 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Link } from 'gatsby';
 
-import Item from './components/Item';
+import List from './components/List';
 
-import './Posts.scss';
 import styles from './Posts.module.scss';
 
 const Posts = () => {
-  let [posts, setPosts] = useState(null);
-  const defaultPosts = [
-    {
-      date: 'Nov 06',
-      title: '46 labs case study: telecom platform redesign',
-      link: '/blog',
-      image: '#',
-    },
-    {
-      date: 'Nov 01',
-      title:
-        'Halo Lab Receives Top UX/UI Design Agency Award in Ukraine By Clutch',
-      link: '/blog',
-      image: '#',
-    },
-    {
-      date: 'Sep 30',
-      title: 'Mobile First Design: it’s worth your attention',
-      link: '/blog',
-      image: '#',
-    },
-  ];
-
-  useEffect(() => {
-    setPosts(defaultPosts);
-  }, []);
-
   return (
     <section className={styles.container}>
       <div className={styles.info}>
@@ -45,15 +17,7 @@ const Posts = () => {
         </Link>
       </div>
 
-      <div className="section__posts-block">
-        {posts ? (
-          posts.map(item => {
-            return <Item key={item.title} {...item} />;
-          })
-        ) : (
-          <div>Loading ...</div>
-        )}
-      </div>
+      <List />
     </section>
   );
 };
