--- src/scenes/Home/components/Posts/components/Block/Block.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import styles from './Block.module.scss';
-
-const Block = () => {
-  return (
-    <div className={styles.container}>
-      <div></div>
-    </div>
-  );
-};
-
-export default Block;
