--- src/scenes/Home/components/Posts/components/Block/Block.module.scss
@@ -1,21 +1,0 @@
-@import '@styles/variables/all';
-@import '@styles/mixins/all';
-
-.container {
-  position: relative;
-  flex-direction: column;
-  display: flex;
-  width: 100%;
-  max-width: 51.43rem;
-  margin-left: 3.33rem;
-
-  @include media-breakpoint-down(md) {
-    max-width: 34.2rem;
-  }
-
-  @include media-breakpoint-down(sm) {
-    margin-top: 6.67rem;
-    margin-left: 0;
-    max-width: 42.69rem;
-  }
-}
