--- src/scenes/Home/components/Posts/components/Block/index.js
@@ -1,3 +1,0 @@
-import Block from './Block';
-
-export default Block;
