--- src/scenes/Home/components/Posts/components/Item/index.js
@@ -1,3 +1,0 @@
-import Item from './Item';
-
-export default Item;
