+++ src/scenes/Home/components/Posts/components/List/index.js
@@ -1,0 +1,3 @@
+import List from './List';
+
+export default List;
